+++ portfolio/src/main/webapp/script.js
@@ -137,3 +137,13 @@ function deleteSpecificComment(comment) {
   fetch('/delete-given-comment', {method: 'POST', body: params});
 }
 
+/** Tells the server to add a comment. */
+async function addComment() {
+  const params = new URLSearchParams();
+  const name = document.getElementById('name').innerHTML;
+  const text = document.getElementById('text').innerHTML;
+  params.append('name', name);
+  params.append('text', text)	
+  const add = await fetch('add-comment', {method: 'POST', body: params});
+  getCommentSection();
+}
